Move BdmsAuthProvider to common/auth
